GO2AllProbes,DatPkg-method now only queries for GO IDs that exist
In making Category/GOstats compatible with db-based ann pkgs, we
encountered some problems with AnnotationDbi raising an error when
asking for a key that doesn't exist.  We plan to fix the env-API to Do
What Environments Do, but in this case it was easy to have the code
simply not ask for things that won't be there.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Category@26489 bc3139a8-67e5-0310-9ffc-ced21a209358